Added image encoding + building Minor bugfixes in error logging Minor GUI changes

+++ Text2Rgb/Text2Rgb/FileHelper.cs
@@ -14,15 +14,21 @@ namespace Text2Rgb
             string errorPath = $"{Environment.CurrentDirectory}\\logs\\error.txt";
             string prefix = $"[{DateTime.Now.ToShortDateString()}] at [{DateTime.Now.ToShortTimeString()}]";
 
-            if (!File.Exists(errorPath))
+            if (!Directory.Exists(errorPath))
             {
-                File.Create(errorPath);
+                Directory.CreateDirectory(errorPath.Replace("error.txt", ""));
+                if (!File.Exists(errorPath))
+                {
+                    File.Create(errorPath);
+                }
             }
+           
 
             //Log the error
             using (StreamWriter sw = File.AppendText(errorPath))
             {
-                sw.WriteLine($"New error logged on {prefix}\r\nError Message{ex.Message}\r\nException Details:\r\n{ex.InnerException}");
+                sw.WriteLine($"New error logged on {prefix}\r\nError Message{ex.Message}\r\nException Details:{Environment.NewLine + ex.InnerException}");
+                sw.WriteLine($"Stack Trace: {ex.StackTrace}");
                 if (!string.IsNullOrWhiteSpace(customMessage))
                 {
                     sw.WriteLine($"Custom developer message: {customMessage}\r\n");
